backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1
python-dotenv==1.0.0

# LangChain and AI packages - compatible versions for structured output
langchain==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.8
pydantic==2.10.5
pydantic-core==2.27.2

# Optional but recommended
langgraph==0.2.55

# Resume processing
PyPDF2==3.0.1
